Fix installation when zlib1g is already installed
No matter what I tried, apt-get wouldn't run zlib1-ng's preinst before
removing zlib1g. This meant that apt-get would fail, because it could
not find libz.so.1.

Now we don't ship a libz.so.1 symlink at all. If no zlib version was
installed previously we create the link in preinst. In prerm we change
the symlink to point at the stock zlib to prevent issues. Triggering
ldconfig takes care of making sure that the symlink is correctly
installed or removed in other cases.